+++ test/Audit.EntityFramework.Full.UnitTest/EfTestsCodeFirst.cs
@@ -106,7 +106,9 @@ namespace Audit.EntityFramework.Full.UnitTest
                 {
                     var validations = ex.EntityValidationErrors.ToList();
                     Assert.AreEqual(1, validations.Count);
+#if !NETCOREAPP3_0
                     Assert.AreEqual(longString, ((dynamic)validations[0].Entry.Entity).Title);
+#endif
                 }
             }